+++ src/FunFair.BuildCheck.Runner/Setup.cs
@@ -216,7 +216,9 @@ internal static class Setup
             .AddProjectCheck<UsingXUnitV3ExtensibilityMustIncludeAnalyzer>()
             .AddProjectCheck<XUnitV3ProjectsShouldDefineTestingPlatformDotnetTestSupport>()
             .AddProjectCheck<XUnitV3ProjectsShouldDefineUseMicrosoftTestingPlatformRunner>()
-            .AddProjectCheck<XUnitV3ProjectsShouldBeAnExecutable>();
+            .AddProjectCheck<XUnitV3ProjectsShouldBeAnExecutable>()
+            .AddProjectCheck<NonTestProjectsMustNotDefineTestingPlatformDotnetTestSupport>()
+            .AddProjectCheck<NonTestProjectsMustNotDefineIsTestingPlatformApplication>();
     }
 
     private static IServiceCollection AddProjectCheck<