+++ CHANGELOG.md
@@ -14,6 +14,8 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 - Added tests for FunFair.BuildCheck.ProjectChecks project checks
 - MustNotUseInternalsVisibleTo rule: fail the build whenever InternalsVisibleTo is present in any .csproj file
 - Additional unit tests to raise FunFair.BuildCheck.Runner code coverage to 95.7%
+- NonTestProjectsMustNotDefineTestingPlatformDotnetTestSupport: errors when a non-test project defines TestingPlatformDotnetTestSupport
+- NonTestProjectsMustNotDefineIsTestingPlatformApplication: errors when a non-test project defines IsTestingPlatformApplication
 ### Fixed
 ### Changed
 - Exclude source-generated code (GeneratedCode attribute) from coverage measurements